+++ ocr_wrapper/google_document_ocr_checkbox_detector.py
@@ -1,11 +1,13 @@
 from __future__ import annotations
 
+import functools
+import os
 from typing import Optional, Union
+
+from PIL import Image
+
 import ocr_wrapper
 from ocr_wrapper import BBox
-import os
-from PIL import Image
-import functools
 
 try:
     from google.api_core.client_options import ClientOptions
@@ -28,7 +30,9 @@ def requires_gcloud(fn):
     return wrapper_decocator
 
 
-def _val_or_env(val: Optional[str], env: str, default: Optional[str] = None) -> Optional[str]:
+def _val_or_env(
+    val: Optional[str], env: str, default: Optional[str] = None
+) -> Optional[str]:
     """Return val if not None, else return the value of the environment variable env, if that is set, else return default."""
     return val if val is not None else os.getenv(env, default)
 
@@ -43,21 +47,22 @@ def _visual_element_to_bbox(visual_element) -> BBox:
     Returns:
         A BBox object with the bounds of the visual element and the associated text character.
     """
-    breakpoint()
-    style2text = {  # Map the style to a fitting unicode character
+    style2text = {  # Mqap the style to a fitting unicode character
         "filled_checkbox": "☑",
         "unfilled_checkbox": "☐",
     }
 
     vertices = visual_element.layout.bounding_poly.normalized_vertices
     vertices = [value for vertex in vertices for value in (vertex.x, vertex.y)]
+    confidence = visual_element.layout.confidence
     bbox = BBox.from_float_list(vertices)
     bbox.text = style2text[visual_element.type_]
 
-    return bbox
+    return bbox, confidence
 
 
 class GoogleDocumentOcrCheckboxDetector:
+    @requires_gcloud
     def __init__(
         self,
         project_id: Optional[str] = None,
@@ -69,7 +74,9 @@ class GoogleDocumentOcrCheckboxDetector:
         self.location = _val_or_env(location, "GOOGLE_DOC_OCR_LOCATION", default="eu")
         self.processor_id = _val_or_env(processor_id, "GOOGLE_DOC_OCR_PROCESSOR_ID")
         self.processor_version = _val_or_env(
-            processor_version, "GOOGLE_DOC_OCR_PROCESSOR_VERSION", default="pretrained-ocr-v2.0-2023-06-02"
+            processor_version,
+            "GOOGLE_DOC_OCR_PROCESSOR_VERSION",
+            default="pretrained-ocr-v2.0-2023-06-02",
         )
 
         self.process_options = documentai.ProcessOptions(
@@ -81,7 +88,9 @@ class GoogleDocumentOcrCheckboxDetector:
         )
 
         self.client = documentai.DocumentProcessorServiceClient(
-            client_options=ClientOptions(api_endpoint=f"{self.location}-documentai.googleapis.com")
+            client_options=ClientOptions(
+                api_endpoint=f"{self.location}-documentai.googleapis.com"
+            )
         )
         if (
             self.project_id is None
@@ -96,14 +105,23 @@ class GoogleDocumentOcrCheckboxDetector:
             self.project_id, self.location, self.processor_id, self.processor_version
         )
 
-    def detect_checkboxes(self, page: Union[Image.Image, documentai.RawDocument]) -> list[BBox]:
+    @requires_gcloud
+    def detect_checkboxes(
+        self, page: Union[Image.Image, documentai.RawDocument]
+    ) -> list[BBox]:
         if isinstance(page, Image.Image):
-            img_byte_arr = ocr_wrapper.OcrWrapper._pil_img_to_compressed(image=page, compression="webp")
-            raw_document = documentai.RawDocument(content=img_byte_arr, mime_type="image/webp")
+            img_byte_arr = ocr_wrapper.OcrWrapper._pil_img_to_compressed(
+                image=page, compression="webp"
+            )
+            raw_document = documentai.RawDocument(
+                content=img_byte_arr, mime_type="image/webp"
+            )
         elif isinstance(page, documentai.RawDocument):
             raw_document = page
         else:
-            raise ValueError("page should be of type Image.Image or documentai.types.RawDocument")
+            raise ValueError(
+                "page should be of type Image.Image or documentai.types.RawDocument"
+            )
 
         # Execute the request with exponential backoff and retry
         request = documentai.ProcessRequest(
@@ -114,6 +132,13 @@ class GoogleDocumentOcrCheckboxDetector:
         result = self.client.process_document(request=request)
 
         result = [
-            _visual_element_to_bbox(visual_element) for visual_element in result.document.pages[0].visual_elements
+            _visual_element_to_bbox(visual_element)
+            for visual_element in result.document.pages[0].visual_elements
         ]
-        return result
+        # For some reason, the system generally returns exactly the same checkbox twice, so we have to get rid of the duplicates
+        result = list(set(result))
+
+        # Separate bboxes and confidence tuples
+        bboxes, confidences = zip(*result)
+
+        return bboxes, confidences